Project-Finished
The goal of this project is to predict the gender of a tennis player from the match stats of 2013 Grand Slam Tournaments, and along the way showcase the Python skills learned during this course.

In order to do so we are going to use the a collection of datasets from the UCI database with the grand slam matches statistics for 2013. There are 8 datasets, one per Grand Slam and Gender.